Rfc: 메인페이지에 지도 보여주기
