+++ src/app/(main)/layout.tsx
@@ -1,4 +1,4 @@
-import "./globals.css";
+import "../globals.css";
 
 export default function RootLayout({
   children,
