--- src/app/page.tsx
@@ -1,12 +1,0 @@
-import Link from "next/link";
-
-const MainPage = () => {
-  return (
-    <>
-      <p>Welcome! Wooddk</p>
-      <Link href="/map">지도페이지</Link>
-    </>
-  );
-};
-
-export default MainPage;
